Divide core lib into header/body
