+++ test/test_port.c
@@ -1,5 +1,5 @@
 #include "unity.h"
-#include "../rackit/rework.c"
+#include "../rackit/core.c"
 
 void test_Port_size(void) {
     // Size is dominated by fallback/output buffer
